expectedStatics
